coin-retrieval-engine/requirements.txt
torch
torchvision
faiss-cpu
opencv-python
fastapi
uvicorn
streamlit
pytest
numpy
scikit-learn
pyyaml
pillow
